move `ManagedCache` and `ManagedFeedImage` to their own files
