+++ EssentialFeed/Feed Cache/CoreData/CoreDataFeedStore.swift
@@ -74,41 +74,3 @@ public class CoreDataFeedStore: FeedStore {
     }
 
 }
-
-@objc(ManagedCache)
-final class ManagedCache: NSManagedObject {
-    @NSManaged var timestamp: Date
-    @NSManaged var feed: NSOrderedSet
-
-    var locals: [LocalFeedImage] {
-        return feed.compactMap {
-            ($0 as? ManagedFeedImage).map {
-                LocalFeedImage(id: $0.id, description: $0.imageDescription, location: $0.location, url: $0.url)
-            }
-        }
-    }
-}
-
-@objc(ManagedFeedImage)
-final class ManagedFeedImage: NSManagedObject {
-    @NSManaged var id: UUID
-    @NSManaged var imageDescription: String?
-    @NSManaged var location: String?
-    @NSManaged var url: URL
-    @NSManaged var cache: ManagedCache
-
-    private convenience init(local: LocalFeedImage, context: NSManagedObjectContext) {
-        self.init(context: context)
-
-        id = local.id
-        imageDescription = local.description
-        location = local.location
-        url = local.url
-    }
-
-    static func build(with images: [LocalFeedImage], in context: NSManagedObjectContext) -> NSOrderedSet {
-        return NSOrderedSet(array: images.map { localFeedImage in
-            return ManagedFeedImage(local: localFeedImage, context: context)
-        })
-    }
-}
